backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
flask-swagger-ui==4.11.1
flask-limiter==3.5.0
psycopg2-binary==2.9.9
werkzeug==2.3.7
python-dotenv==1.0.0
flask-cors==4.0.0 